refactor(app): extract message actions
